moodbot-api/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
nltk==3.9.2
numpy==2.3.5
packaging==25.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2025.11.3
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
Werkzeug==3.1.3
deep-translator==1.11.4
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
